requirements.txt
Django==5.1
psycopg2-binary==2.9.9
pillow==10.4.0
freezegun==1.5.1
django-solo==2.3.0
requests==2.32.3
timeout-decorator==0.5.0
